+++ better-me-api/Controllers/UsersController.cs
@@ -8,6 +8,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using BetterMeApi.Models;
 using BetterMeApi.Repositories;
+using System.Security.Claims;
 
 namespace BetterMeApi.Controllers
 {
@@ -22,7 +23,6 @@ namespace BetterMeApi.Controllers
             _userRepository = userRepository;
         }
 
-        // GET api/values
         [HttpGet]
         public IActionResult Get([FromQuery(), RegularExpression(@"^(?("")("".+?(?<!\\)""@)|(([0-9a-z]((\.(?!\.))|[-!#\$%&'\*\+/=\?\^`\{\}\|~\w])*)(?<=[0-9a-z])@))(?(\[)(\[(\d{1,3}\.){3}\d{1,3}\])|(([0-9a-z][-\w]*[0-9a-z]*\.)+[a-z0-9][\-a-z0-9]{0,22}[a-z0-9]))$")] string email)
         {
@@ -30,29 +30,28 @@ namespace BetterMeApi.Controllers
             {
                 if (!ModelState.IsValid)
                 {
-                    return BadRequest();
+                    return BadRequest(ErrorCode.DataProvidedIsInvalid.ToString());
                 }
 
-                return Ok(_userRepository.All.Where(user => user.Email == email));
+                return Ok(_userRepository.Query(email));
             }
 
             return Ok(_userRepository.All);
         }
 
-        // GET api/values/5
         [HttpGet("{id}")]
         public IActionResult Get(long id)
         {
             var item = _userRepository.Find(id);
             if (item == null)
             {
-                return NotFound(ErrorCode.RecordNotFound.ToString());
+                return NotFound(ErrorCode.ItemNotFound.ToString());
             }
             
             return Ok(item);
         }
 
-        // POST api/values
+        /*
         [HttpPost]
         public IActionResult Post([FromBody]User item)
         {
@@ -60,12 +59,12 @@ namespace BetterMeApi.Controllers
             {
                 if (item == null || !ModelState.IsValid || item.UserId != 0)
                 {
-                    return BadRequest(ErrorCode.ItemNameAndNotesRequired.ToString());
+                    return BadRequest(ErrorCode.DataProvidedIsInvalid.ToString());
                 }
 
                 if (_userRepository.DoesItemExist(item.UserId) || _userRepository.DoesItemExist(item.Email))
                 {
-                    return StatusCode(StatusCodes.Status409Conflict, ErrorCode.ItemIDInUse.ToString());
+                    return StatusCode(StatusCodes.Status409Conflict, ErrorCode.ItemAlreadyExists.ToString());
                 }
                 _userRepository.Insert(item);
             }
@@ -76,8 +75,8 @@ namespace BetterMeApi.Controllers
             
             return Ok(item);
         }
+        */
 
-        // PUT api/values/5
         [HttpPut("{id}")]
         public IActionResult Put([FromBody] User item)
         {
@@ -85,13 +84,15 @@ namespace BetterMeApi.Controllers
             {
                 if (item == null || !ModelState.IsValid)
                 {
-                    return BadRequest(ErrorCode.ItemNameAndNotesRequired.ToString());
+                    return BadRequest(ErrorCode.DataProvidedIsInvalid.ToString());
                 }
-                var existingItem = _userRepository.Find(item.UserId);
-                if (existingItem == null)
+
+                var userItem = _userRepository.Find(item.UserId);
+                if (userItem == null || userItem.Email != User.FindFirstValue(ClaimTypes.Email))
                 {
-                    return NotFound(ErrorCode.RecordNotFound.ToString());
+                    return NotFound(ErrorCode.ItemNotFound.ToString());
                 }
+
                 _userRepository.Update(item);
             }
             catch (Exception)
@@ -101,16 +102,15 @@ namespace BetterMeApi.Controllers
             return NoContent();
         }
 
-        // DELETE api/values/5
         [HttpDelete("{id}")]
         public IActionResult Delete(long id)
         {
             try
             {
                 var item = _userRepository.Find(id);
-                if (item == null)
+                if (item == null || item.Email != User.FindFirstValue(ClaimTypes.Email))
                 {
-                    return NotFound(ErrorCode.RecordNotFound.ToString());
+                    return NotFound(ErrorCode.ItemNotFound.ToString());
                 }
                 _userRepository.Delete(id);
             }
